+++ client/beadledom-client-example/example-client/src/main/java/com/cerner/beadledom/client/example/client/ExampleOneClientModule.java
@@ -54,9 +54,9 @@ public class ExampleOneClientModule extends AbstractModule {
   BeadledomClientConfiguration provideClientConfig() {
     return BeadledomClientConfiguration.builder()
         .maxPooledPerRouteSize(60)
-        .socketTimeoutMillis(60)
-        .connectionTimeoutMillis(60)
-        .ttlMillis(60)
+        .socketTimeoutMillis(200)
+        .connectionTimeoutMillis(200)
+        .ttlMillis(200)
         .connectionPoolSize(60).build();
   }
 }